voice_to_text_app/requirements.txt
whisper
pyannote.audio
pydub
torch
torchaudio
ffmpeg-python
